Compile fixes

+++ xen/runtime/include/inttypes.h
@@ -252,8 +252,8 @@ typedef unsigned int uint32_t __attribute__((__mode__(__SI__)));
 typedef unsigned int uint32_t;
 #endif
 
-typedef signed long int intptr_t;
-typedef unsigned long int uintptr_t;
+typedef signed int intptr_t;
+typedef unsigned int uintptr_t;
 
 #if __WORDSIZE == 64
 typedef signed long int64_t;
